refactor: Move files for Setup fsd
