--- src/components/AddTodo/models/index.ts
@@ -1,1 +1,0 @@
-export * from './useAddTodoState';
